APM: fixed some build warnings and type errors

+++ ArduPlane/commands.pde
@@ -85,7 +85,7 @@ static struct Location get_cmd_with_index(int i)
 static void set_cmd_with_index(struct Location temp, int i)
 {
 	i = constrain(i, 0, g.command_total.get());
-	uint32_t mem = WP_START_BYTE + (i * WP_SIZE);
+	intptr_t mem = WP_START_BYTE + (i * WP_SIZE);
 
 	// Set altitude options bitmask
 	// XXX What is this trying to do?
